requirements.txt
circlify
numpy
# opencv-python
pylabels
reportlab